requirements.txt
Django==4.0.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.1.1
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2

django-filter~=21.1